+++ src/test/kotlin/io/kjson/yaml/YAMLTest.kt
@@ -2,7 +2,7 @@
  * @(#) YAMLTest.kt
  *
  * kjson-yaml  Kotlin YAML processor
- * Copyright (c) 2020, 2021 Peter Wall
+ * Copyright (c) 2020, 2021, 2023 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -43,6 +43,16 @@ class YAMLTest {
         expect("data") { YAML.parse(inputStream).rootNode.asObject["key"].asString }
         val reader = file.reader()
         expect("data") { YAML.parse(reader).rootNode.asObject["key"].asString }
+    }
+
+    @Test fun `should use YAML object functions for multi-document streams`() {
+        val file = File("src/test/resources/keyblock.yaml")
+        val result = YAML.parseStream(file)
+        expect("data") { result.single().rootNode.asObject["key"].asString }
+        val inputStream = file.inputStream()
+        expect("data") { YAML.parseStream(inputStream).single().rootNode.asObject["key"].asString }
+        val reader = file.reader()
+        expect("data") { YAML.parseStream(reader).single().rootNode.asObject["key"].asString }
     }
 
 }